+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/filler/impl/SubqueryFiller.java
@@ -19,6 +19,9 @@ package io.shardingsphere.core.parsing.antlr.filler.impl;
 
 import io.shardingsphere.core.metadata.table.ShardingTableMetaData;
 import io.shardingsphere.core.parsing.antlr.filler.SQLStatementFiller;
+import io.shardingsphere.core.parsing.antlr.filler.impl.dql.GroupByFiller;
+import io.shardingsphere.core.parsing.antlr.filler.impl.dql.OrderByFiller;
+import io.shardingsphere.core.parsing.antlr.filler.impl.dql.SelectClauseFiller;
 import io.shardingsphere.core.parsing.antlr.sql.segment.expr.SubquerySegment;
 import io.shardingsphere.core.parsing.parser.sql.SQLStatement;
 import io.shardingsphere.core.parsing.parser.sql.dql.select.SelectStatement;
