+++ sharding-core/src/main/java/io/shardingsphere/core/parsing/antlr/filler/impl/dql/OrderByFiller.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2016-2018 shardingsphere.io.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * </p>
+ */
+
+package io.shardingsphere.core.parsing.antlr.filler.impl.dql;
+
+import io.shardingsphere.core.metadata.table.ShardingTableMetaData;
+import io.shardingsphere.core.parsing.antlr.filler.SQLStatementFiller;
+import io.shardingsphere.core.parsing.antlr.sql.segment.order.OrderBySegment;
+import io.shardingsphere.core.parsing.antlr.sql.segment.order.item.OrderByItemSegment;
+import io.shardingsphere.core.parsing.parser.sql.SQLStatement;
+import io.shardingsphere.core.parsing.parser.sql.dql.select.SelectStatement;
+import io.shardingsphere.core.rule.ShardingRule;
+
+/**
+ * Order by filler.
+ *
+ * @author duhongjun
+ */
+public final class OrderByFiller implements SQLStatementFiller<OrderBySegment> {
+    
+    @Override
+    public void fill(final OrderBySegment sqlSegment, final SQLStatement sqlStatement, final String sql, final ShardingRule shardingRule, final ShardingTableMetaData shardingTableMetaData) {
+        SelectStatement selectStatement = (SelectStatement) sqlStatement;
+        if (!selectStatement.getSubQueryStatements().isEmpty()) {
+            return;
+        }
+        for (OrderByItemSegment each : sqlSegment.getOrderByItems()) {
+            selectStatement.getOrderByItems().add(new OrderItemBuilder(selectStatement, each).createOrderItem());
+        }
+    }
+}
